refactor: move calculate_mining_probability to validations crate
