+++ node/src/actors/chain_manager/mining.rs
@@ -20,7 +20,7 @@ use witnet_data_structures::{
     chain::{
         Block, BlockHeader, BlockMerkleRoots, BlockTransactions, Bn256PublicKey, CheckpointBeacon,
         CheckpointVRF, DataRequestOutput, EpochConstants, Hash, Hashable, Input, PublicKeyHash,
-        ReputationEngine, TransactionsPool, ValueTransferOutput,
+        TransactionsPool, ValueTransferOutput,
     },
     data_request::{
         calculate_witness_reward, calculate_witness_reward_before_second_hard_fork, create_tally,
@@ -42,9 +42,10 @@ use witnet_futures_utils::TryFutureExt2;
 use witnet_rad::{error::RadError, types::serial_iter_decode};
 use witnet_util::timestamp::get_timestamp;
 use witnet_validations::validations::{
-    block_reward, calculate_liars_and_errors_count_from_tally, calculate_randpoe_threshold,
-    calculate_reppoe_threshold, dr_transaction_fee, merkle_tree_root, radon_report_from_error,
-    tally_bytes_on_encode_error, update_utxo_diff, vt_transaction_fee, Wit,
+    block_reward, calculate_liars_and_errors_count_from_tally, calculate_mining_probability,
+    calculate_randpoe_threshold, calculate_reppoe_threshold, dr_transaction_fee, merkle_tree_root,
+    radon_report_from_error, tally_bytes_on_encode_error, update_utxo_diff, vt_transaction_fee,
+    Wit,
 };
 
 use crate::{
@@ -995,95 +996,9 @@ pub fn build_block(
     (block_header, txns)
 }
 
-#[allow(clippy::many_single_char_names)]
-fn internal_calculate_mining_probability(
-    rf: u32,
-    n: f64,
-    k: u32, // k: iterative rf until reach bf
-    m: i32, // M: nodes with reputation greater than me
-    l: i32, // L: nodes with reputation equal than me
-    r: i32, // R: nodes with reputation less than me
-) -> f64 {
-    if k == rf {
-        let rf = f64::from(rf);
-        // Prob to mine is the probability that a node with the same reputation than me mine,
-        // divided by all the nodes with the same reputation:
-        // 1/L * (1 - ((N-RF)/N)^L)
-        let prob_to_mine = (1.0 / f64::from(l)) * (1.0 - ((n - rf) / n).powi(l));
-        // Prob that a node with more reputation than me mine is:
-        // ((N-RF)/N)^M
-        let prob_greater_neg = ((n - rf) / n).powi(m);
-
-        prob_to_mine * prob_greater_neg
-    } else {
-        let k = f64::from(k);
-        // Here we take into account that rf = 1 because is only a new slot
-        let prob_to_mine = (1.0 / f64::from(l)) * (1.0 - ((n - 1.0) / n).powi(l));
-        // The same equation than before
-        let prob_bigger_neg = ((n - k) / n).powi(m);
-        // Prob that a node with less or equal reputation than me mine with a lower slot is:
-        // ((N+1-RF)/N)^(L+R-1)
-        let prob_lower_slot_neg = ((n + 1.0 - k) / n).powi(l + r - 1);
-
-        prob_to_mine * prob_bigger_neg * prob_lower_slot_neg
-    }
-}
-
-fn calculate_mining_probability(
-    rep_engine: &ReputationEngine,
-    own_pkh: PublicKeyHash,
-    rf: u32,
-    bf: u32,
-) -> f64 {
-    let n = u32::try_from(rep_engine.ars().active_identities_number()).unwrap();
-
-    // In case of any active node, the probability is maximum
-    if n == 0 {
-        return 1.0;
-    }
-
-    // First we need to know how many nodes have more or equal reputation than us
-    let own_rep = rep_engine.trs().get(&own_pkh);
-    let is_active_node = rep_engine.ars().contains(&own_pkh);
-    let mut greater = 0;
-    let mut equal = 0;
-    let mut less = 0;
-    for &active_id in rep_engine.ars().active_identities() {
-        let rep = rep_engine.trs().get(&active_id);
-        match (rep.0 > 0, own_rep.0 > 0) {
-            (true, false) => greater += 1,
-            (false, true) => less += 1,
-            _ => equal += 1,
-        }
-    }
-    // In case of not being active, the equal value is plus 1.
-    if !is_active_node {
-        equal += 1;
-    }
-
-    if rf > n && greater == 0 {
-        // In case of replication factor exceed the active node number and being the most reputed
-        // we obtain the maximum probability divided in the nodes we share the same reputation
-        1.0 / f64::from(equal)
-    } else if rf > n && greater > 0 {
-        // In case of replication factor exceed the active node number and not being the most reputed
-        // we obtain the minimum probability
-        0.0
-    } else {
-        let mut aux =
-            internal_calculate_mining_probability(rf, f64::from(n), rf, greater, equal, less);
-        let mut k = rf + 1;
-        while k <= bf && k <= n {
-            aux += internal_calculate_mining_probability(rf, f64::from(n), k, greater, equal, less);
-            k += 1;
-        }
-        aux
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use std::convert::{TryFrom, TryInto};
+    use std::convert::TryInto;
 
     use secp256k1::{
         PublicKey as Secp256k1_PublicKey, Secp256k1, SecretKey as Secp256k1_SecretKey,
@@ -1658,234 +1573,5 @@ mod tests {
         assert_eq!(public_key, public_key2);
 
         assert!(verify(secp, &public_key2, &data, &signature2).is_ok());
-    }
-
-    fn init_rep_engine(v_rep: Vec<u32>) -> (ReputationEngine, Vec<PublicKeyHash>) {
-        let mut rep_engine = ReputationEngine::new(1000);
-
-        let mut ids = vec![];
-        for (i, &rep) in v_rep.iter().enumerate() {
-            let pkh = PublicKeyHash::from_bytes(&[u8::try_from(i).unwrap(); 20]).unwrap();
-            rep_engine
-                .trs_mut()
-                .gain(Alpha(10), vec![(pkh, Reputation(rep))])
-                .unwrap();
-            ids.push(pkh);
-        }
-        rep_engine.ars_mut().push_activity(ids.clone());
-
-        (rep_engine, ids)
-    }
-
-    fn calculate_mining_probs(v_rep: Vec<u32>, rf: u32, bf: u32) -> (Vec<f64>, f64) {
-        let v_rep_len = v_rep.len();
-        let (rep_engine, ids) = init_rep_engine(v_rep);
-        let n = rep_engine.ars().active_identities_number();
-        assert_eq!(n, v_rep_len);
-        assert_eq!(ids.len(), v_rep_len);
-
-        let mut probs = vec![];
-        for id in ids {
-            probs.push(calculate_mining_probability(&rep_engine, id, rf, bf))
-        }
-
-        let new_pkh = PublicKeyHash::from_bytes(&[0xFF; 20]).unwrap();
-        let new_prob = calculate_mining_probability(&rep_engine, new_pkh, rf, bf);
-
-        (probs, new_prob)
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-    #[test]
-    fn calculate_mining_probabilities_rf1_bf1() {
-        let v_rep = vec![10, 8, 8, 8, 5, 5, 5, 5, 0, 0];
-        let (probs, new_prob) = calculate_mining_probs(v_rep, 1, 1);
-
-        for &prob in &probs[0..8] {
-            assert_eq!(
-                (prob * 10_000.0).round() as u32,
-                (7.12_f64 * 100.0).round() as u32
-            );
-        }
-
-        for &prob in &probs[8..10] {
-            assert_eq!(
-                (prob * 10_000.0).round() as u32,
-                (4.09_f64 * 100.0).round() as u32
-            );
-        }
-
-        assert_eq!(
-            (new_prob * 10_000.0).round() as u32,
-            (3.89_f64 * 100.0).round() as u32
-        );
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-    #[test]
-    fn calculate_mining_probabilities_rf1_bf2() {
-        let v_rep = vec![10, 8, 8, 8, 5, 5, 5, 5, 2, 2];
-        let (probs, new_prob) = calculate_mining_probs(v_rep, 1, 2);
-
-        for prob in probs {
-            assert_eq!(
-                (prob * 10_000.0).round() as u32,
-                (9.04_f64 * 100.0).round() as u32
-            );
-        }
-
-        assert_eq!(
-            (new_prob * 10_000.0).round() as u32,
-            (4.56_f64 * 100.0).round() as u32
-        );
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-    #[test]
-    fn calculate_mining_probabilities_rf2_bf2() {
-        let v_rep = vec![10, 8, 8, 8, 5, 5, 5, 5, 2, 2];
-        let (probs, new_prob) = calculate_mining_probs(v_rep, 2, 2);
-
-        for prob in probs {
-            assert_eq!(
-                (prob * 10_000.0).round() as u32,
-                (8.93_f64 * 100.0).round() as u32
-            );
-        }
-
-        assert_eq!(
-            (new_prob * 10_000.0).round() as u32,
-            (2.15_f64 * 100.0).round() as u32
-        );
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-    #[test]
-    fn calculate_mining_probabilities_rf4_bf8() {
-        let v_rep = vec![10, 8, 8, 8, 5, 5, 5, 5, 2, 2];
-        let (probs, new_prob) = calculate_mining_probs(v_rep, 4, 8);
-
-        for prob in probs {
-            assert_eq!(
-                (prob * 10_000.0).round() as u32,
-                (10.02_f64 * 100.0).round() as u32
-            );
-        }
-
-        assert_eq!(
-            (new_prob * 10_000.0).round() as u32,
-            (0.25_f64 * 100.0).round() as u32
-        );
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-    #[test]
-    fn calculate_mining_probabilities_rf1_bf1_10() {
-        let v_rep = vec![10; 10];
-        let (probs, new_prob) = calculate_mining_probs(v_rep, 1, 1);
-
-        assert_eq!(
-            (probs[0] * 10_000.0).round() as u32,
-            (6.51_f64 * 100.0).round() as u32
-        );
-
-        assert_eq!(
-            (new_prob * 10_000.0).round() as u32,
-            (3.49_f64 * 100.0).round() as u32
-        );
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-    #[test]
-    fn calculate_mining_probabilities_rf1_bf1_100() {
-        let v_rep = vec![10; 100];
-        let (probs, new_prob) = calculate_mining_probs(v_rep, 1, 1);
-
-        assert_eq!(
-            (probs[0] * 10_000.0).round() as u32,
-            (0.63_f64 * 100.0).round() as u32
-        );
-
-        assert_eq!(
-            (new_prob * 10_000.0).round() as u32,
-            (0.37_f64 * 100.0).round() as u32
-        );
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-    #[test]
-    fn calculate_mining_probabilities_rf4_bf8_100() {
-        let v_rep = vec![10; 100];
-        let (probs, new_prob) = calculate_mining_probs(v_rep, 4, 8);
-
-        assert_eq!(
-            (probs[0] * 10_000.0).round() as u32,
-            (1.0_f64 * 100.0).round() as u32
-        );
-
-        assert_eq!(
-            (new_prob * 10_000.0).round() as u32,
-            (0.08_f64 * 100.0).round() as u32
-        );
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-    #[test]
-    fn calculate_mining_probabilities_rf4_bf8_100_diff() {
-        let mut v_rep = vec![10; 25];
-        v_rep.extend(vec![8; 25]);
-        v_rep.extend(vec![6; 25]);
-        v_rep.extend(vec![4; 25]);
-        let (probs, new_prob) = calculate_mining_probs(v_rep, 4, 8);
-
-        assert_eq!(
-            (probs[0] * 10_000.0).round() as u32,
-            (1_f64 * 100.0).round() as u32
-        );
-        assert_eq!(
-            (probs[25] * 10_000.0).round() as u32,
-            (1_f64 * 100.0).round() as u32
-        );
-        assert_eq!(
-            (probs[50] * 10_000.0).round() as u32,
-            (1_f64 * 100.0).round() as u32
-        );
-        assert_eq!(
-            (probs[75] * 10_000.0).round() as u32,
-            (1_f64 * 100.0).round() as u32
-        );
-
-        assert_eq!(
-            (new_prob * 10_000.0).round() as u32,
-            (0.08_f64 * 100.0).round() as u32
-        );
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-    #[test]
-    fn calculate_mining_probabilities_rf_high() {
-        let v_rep = vec![10, 8, 8, 2];
-        let (probs, new_prob) = calculate_mining_probs(v_rep, 4, 8);
-
-        for prob in probs {
-            assert_eq!(
-                (prob * 10_000.0).round() as u32,
-                (25_f64 * 100.0).round() as u32
-            );
-        }
-
-        assert_eq!(
-            (new_prob * 10_000.0).round() as u32,
-            (0_f64 * 100.0).round() as u32
-        );
     }
 }
