+++ validations/src/tests/randpoe.rs
@@ -1,6 +1,7 @@
 use approx::assert_abs_diff_eq;
+use std::convert::TryFrom;
 use witnet_data_structures::{
-    chain::Hash,
+    chain::{Alpha, Hash, PublicKeyHash, Reputation, ReputationEngine},
     mainnet_validations::{ActiveWips, SECOND_HARD_FORK},
 };
 
@@ -198,4 +199,233 @@ fn vrf_sections() {
     assert_eq!(a.slot(&h1), 1);
     assert_eq!(a.slot(&h2), 1);
     assert_eq!(a.slot(&h3), 2);
+}
+
+fn init_rep_engine(v_rep: Vec<u32>) -> (ReputationEngine, Vec<PublicKeyHash>) {
+    let mut rep_engine = ReputationEngine::new(1000);
+
+    let mut ids = vec![];
+    for (i, &rep) in v_rep.iter().enumerate() {
+        let pkh = PublicKeyHash::from_bytes(&[u8::try_from(i).unwrap(); 20]).unwrap();
+        rep_engine
+            .trs_mut()
+            .gain(Alpha(10), vec![(pkh, Reputation(rep))])
+            .unwrap();
+        ids.push(pkh);
+    }
+    rep_engine.ars_mut().push_activity(ids.clone());
+
+    (rep_engine, ids)
+}
+
+fn calculate_mining_probs(v_rep: Vec<u32>, rf: u32, bf: u32) -> (Vec<f64>, f64) {
+    let v_rep_len = v_rep.len();
+    let (rep_engine, ids) = init_rep_engine(v_rep);
+    let n = rep_engine.ars().active_identities_number();
+    assert_eq!(n, v_rep_len);
+    assert_eq!(ids.len(), v_rep_len);
+
+    let mut probs = vec![];
+    for id in ids {
+        probs.push(calculate_mining_probability(&rep_engine, id, rf, bf))
+    }
+
+    let new_pkh = PublicKeyHash::from_bytes(&[0xFF; 20]).unwrap();
+    let new_prob = calculate_mining_probability(&rep_engine, new_pkh, rf, bf);
+
+    (probs, new_prob)
+}
+
+// FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
+#[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
+#[test]
+fn calculate_mining_probabilities_rf1_bf1() {
+    let v_rep = vec![10, 8, 8, 8, 5, 5, 5, 5, 0, 0];
+    let (probs, new_prob) = calculate_mining_probs(v_rep, 1, 1);
+
+    for &prob in &probs[0..8] {
+        assert_eq!(
+            (prob * 10_000.0).round() as u32,
+            (7.12_f64 * 100.0).round() as u32
+        );
+    }
+
+    for &prob in &probs[8..10] {
+        assert_eq!(
+            (prob * 10_000.0).round() as u32,
+            (4.09_f64 * 100.0).round() as u32
+        );
+    }
+
+    assert_eq!(
+        (new_prob * 10_000.0).round() as u32,
+        (3.89_f64 * 100.0).round() as u32
+    );
+}
+
+// FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
+#[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
+#[test]
+fn calculate_mining_probabilities_rf1_bf2() {
+    let v_rep = vec![10, 8, 8, 8, 5, 5, 5, 5, 2, 2];
+    let (probs, new_prob) = calculate_mining_probs(v_rep, 1, 2);
+
+    for prob in probs {
+        assert_eq!(
+            (prob * 10_000.0).round() as u32,
+            (9.04_f64 * 100.0).round() as u32
+        );
+    }
+
+    assert_eq!(
+        (new_prob * 10_000.0).round() as u32,
+        (4.56_f64 * 100.0).round() as u32
+    );
+}
+
+// FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
+#[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
+#[test]
+fn calculate_mining_probabilities_rf2_bf2() {
+    let v_rep = vec![10, 8, 8, 8, 5, 5, 5, 5, 2, 2];
+    let (probs, new_prob) = calculate_mining_probs(v_rep, 2, 2);
+
+    for prob in probs {
+        assert_eq!(
+            (prob * 10_000.0).round() as u32,
+            (8.93_f64 * 100.0).round() as u32
+        );
+    }
+
+    assert_eq!(
+        (new_prob * 10_000.0).round() as u32,
+        (2.15_f64 * 100.0).round() as u32
+    );
+}
+
+// FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
+#[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
+#[test]
+fn calculate_mining_probabilities_rf4_bf8() {
+    let v_rep = vec![10, 8, 8, 8, 5, 5, 5, 5, 2, 2];
+    let (probs, new_prob) = calculate_mining_probs(v_rep, 4, 8);
+
+    for prob in probs {
+        assert_eq!(
+            (prob * 10_000.0).round() as u32,
+            (10.02_f64 * 100.0).round() as u32
+        );
+    }
+
+    assert_eq!(
+        (new_prob * 10_000.0).round() as u32,
+        (0.25_f64 * 100.0).round() as u32
+    );
+}
+
+// FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
+#[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
+#[test]
+fn calculate_mining_probabilities_rf1_bf1_10() {
+    let v_rep = vec![10; 10];
+    let (probs, new_prob) = calculate_mining_probs(v_rep, 1, 1);
+
+    assert_eq!(
+        (probs[0] * 10_000.0).round() as u32,
+        (6.51_f64 * 100.0).round() as u32
+    );
+
+    assert_eq!(
+        (new_prob * 10_000.0).round() as u32,
+        (3.49_f64 * 100.0).round() as u32
+    );
+}
+
+// FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
+#[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
+#[test]
+fn calculate_mining_probabilities_rf1_bf1_100() {
+    let v_rep = vec![10; 100];
+    let (probs, new_prob) = calculate_mining_probs(v_rep, 1, 1);
+
+    assert_eq!(
+        (probs[0] * 10_000.0).round() as u32,
+        (0.63_f64 * 100.0).round() as u32
+    );
+
+    assert_eq!(
+        (new_prob * 10_000.0).round() as u32,
+        (0.37_f64 * 100.0).round() as u32
+    );
+}
+
+// FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
+#[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
+#[test]
+fn calculate_mining_probabilities_rf4_bf8_100() {
+    let v_rep = vec![10; 100];
+    let (probs, new_prob) = calculate_mining_probs(v_rep, 4, 8);
+
+    assert_eq!(
+        (probs[0] * 10_000.0).round() as u32,
+        (1.0_f64 * 100.0).round() as u32
+    );
+
+    assert_eq!(
+        (new_prob * 10_000.0).round() as u32,
+        (0.08_f64 * 100.0).round() as u32
+    );
+}
+
+// FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
+#[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
+#[test]
+fn calculate_mining_probabilities_rf4_bf8_100_diff() {
+    let mut v_rep = vec![10; 25];
+    v_rep.extend(vec![8; 25]);
+    v_rep.extend(vec![6; 25]);
+    v_rep.extend(vec![4; 25]);
+    let (probs, new_prob) = calculate_mining_probs(v_rep, 4, 8);
+
+    assert_eq!(
+        (probs[0] * 10_000.0).round() as u32,
+        (1_f64 * 100.0).round() as u32
+    );
+    assert_eq!(
+        (probs[25] * 10_000.0).round() as u32,
+        (1_f64 * 100.0).round() as u32
+    );
+    assert_eq!(
+        (probs[50] * 10_000.0).round() as u32,
+        (1_f64 * 100.0).round() as u32
+    );
+    assert_eq!(
+        (probs[75] * 10_000.0).round() as u32,
+        (1_f64 * 100.0).round() as u32
+    );
+
+    assert_eq!(
+        (new_prob * 10_000.0).round() as u32,
+        (0.08_f64 * 100.0).round() as u32
+    );
+}
+
+// FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
+#[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
+#[test]
+fn calculate_mining_probabilities_rf_high() {
+    let v_rep = vec![10, 8, 8, 2];
+    let (probs, new_prob) = calculate_mining_probs(v_rep, 4, 8);
+
+    for prob in probs {
+        assert_eq!(
+            (prob * 10_000.0).round() as u32,
+            (25_f64 * 100.0).round() as u32
+        );
+    }
+
+    assert_eq!(
+        (new_prob * 10_000.0).round() as u32,
+        (0_f64 * 100.0).round() as u32
+    );
 }
